Use long double as the type of possibiliy.

+++ include/algorithms/alg_inside_outside_main.h
@@ -8,9 +8,9 @@
 #include "kernels/update_parameters.cuh"
 #include "kernels/expect_count.cuh"
 
-float* outside_algorithm(float* mu, float* beta, const uint32_t* sequence, uint32_t* pretermination_lookuptable, 
-
-                        uint32_t* grammar_index, uint32_t* grammar_table, float* alpha, 
+long double* outside_algorithm(long double* mu, long double* beta, const uint32_t* sequence, 
+    uint32_t* pretermination_lookuptable, 
+                        uint32_t* grammar_index, uint32_t* grammar_table, long double* alpha, 
                         int sequence_length, int n_syms, int N, int T, int MS, int n_grammars,
                         std::vector<std::tuple<uint32_t, uint32_t>> inside_order_1_rule_iteration_path
                         #ifdef DEBUG_INSIDE_ALGORITHM
@@ -19,16 +19,17 @@ float* outside_algorithm(float* mu, float* beta, const uint32_t* sequence, uint3
 );
 
 
-float* em_algorithm_calculate_expection_count(float* count, float* mu, float* beta, const uint32_t* sequence, uint32_t* pretermination_lookuptable, 
-                        uint32_t* grammar_index, uint32_t* grammar_table, float* alpha, 
+long double* em_algorithm_calculate_expection_count(long double* count, long double* mu, long double* beta, 
+                        const uint32_t* sequence, uint32_t* pretermination_lookuptable, 
+                        uint32_t* grammar_index, uint32_t* grammar_table, long double* alpha, 
                         int sequence_length, int n_syms, int N, int T, int MS, int n_grammars
                         #ifdef DEBUG_INSIDE_ALGORITHM
                         ,pcfg* grammar
                         #endif
 );
 
-float* inside_algorithm(const uint32_t* sequence, uint32_t* pretermination_lookuptable, 
-                        uint32_t* grammar_index, uint32_t* grammar_table, float* alpha, 
+long double* inside_algorithm(const uint32_t* sequence, uint32_t* pretermination_lookuptable, 
+                        uint32_t* grammar_index, uint32_t* grammar_table, long double* alpha, 
                         int sequence_length, int n_syms, int N, int T, int MS, int n_grammars,
                         std::vector<std::tuple<uint32_t, uint32_t>> inside_order_1_rule_iteration_path, pcfg* grammar);
