+++ include/kernels/inside.cuh
@@ -25,13 +25,13 @@
 #ifdef USE_CUDA
 __global__
 #endif
-void kernel_inside_alpha_zerolization(float* alpha, int N, int MS);
+void kernel_inside_alpha_zerolization(long double* alpha, int N, int MS);
 #ifdef USE_CUDA
 __global__ 
 #endif
 void kernel_inside_base_fill_alpha(  
         const uint32_t* sequence, uint32_t* pretermination_lookuptable, 
-        uint32_t* grammar_index, uint32_t* grammar_table, float* alpha, 
+        uint32_t* grammar_index, uint32_t* grammar_table, long double* alpha, 
         int sequence_length, int n_syms, int N, int T, int MS, int n_grammars
         #ifdef DEBUG_INSIDE_ALGORITHM
         , pcfg* grammar
@@ -42,7 +42,7 @@ void kernel_inside_base_fill_alpha(
 __global__ 
 #endif
 void kernel_inside_computeSpanKernel(const uint32_t* sequence, uint32_t* pretermination_lookuptable, 
-        uint32_t* grammar_index, uint32_t* grammar_table, float* alpha, 
+        uint32_t* grammar_index, uint32_t* grammar_table, long double* alpha, 
         int sequence_length, int n_syms, int N, int T, int MS, int n_grammars,
         std::vector<std::tuple<uint32_t, uint32_t>> inside_order_1_rule_iteration_path
         #ifdef DEBUG_INSIDE_ALGORITHM
