+++ unit_test/hashtable_test.cpp
@@ -1,8 +1,8 @@
 #define BOOST_TEST_MODULE GRAMMAR_TABLE_UNITEST_Module
 #include <boost/test/included/unit_test.hpp>
 #include "grammar.hpp"
-float test_hashtable(pcfg* grammar) {
-    return 0.0f;
+long double test_hashtable(pcfg* grammar) {
+    return 0.0L;
 }
 
 BOOST_AUTO_TEST_CASE(test_hashtable) {
